prueba de flatmap y map

+++ src/main/java/com/gaspar/pizzati/controller/ProductoController.java
@@ -27,7 +27,7 @@ public class ProductoController {
     }
 
     @GetMapping("/{id}")
-    Producto getProducto(@PathVariable Long id){
+    Producto getProducto(@PathVariable Integer id){
         return service.getProducto(id);
     }
 }
